Test si github est disponible
S'il n'est pas disponible les 'sample schemes' et srvct.bash ne sont pas
téléchargé, l'installation n'est pas stoppée.

+++ database_servers/clone_master.sh
@@ -486,21 +486,27 @@ function configure_oracle_accounts
 
 	if [ $cfg_db_type != fs ]
 	then
-		typeset -r srvctl_script="srvctl_${oracle_release%.*.*}.bash"
-		info "install bash completion for srvctl"
-		typeset -r BACKUP_PWD="$PWD"
-		fake_exec_cmd "cd ~/plescripts/tmp"
-		cd ~/plescripts/tmp
-		exec_cmd rm -f $srvctl_script
-		exec_cmd -c wget https://raw.githubusercontent.com/PhilippeLeroux/oracle_bash_completion/master/$srvctl_script
-		if [ $? -eq 0 ]
+		if ! ping_test github.com
 		then
-			exec_cmd scp	$srvctl_script	\
-							root@$cfg_server_name:/etc/bash_completion.d/
+			warning "github.com not available."
+			LN
+		else
+			typeset -r srvctl_script="srvctl_${oracle_release%.*.*}.bash"
+			info "install bash completion for srvctl"
+			typeset -r BACKUP_PWD="$PWD"
+			fake_exec_cmd "cd ~/plescripts/tmp"
+			cd ~/plescripts/tmp
+			exec_cmd rm -f $srvctl_script
+			exec_cmd -c wget https://raw.githubusercontent.com/PhilippeLeroux/oracle_bash_completion/master/$srvctl_script
+			if [ $? -eq 0 ]
+			then
+				exec_cmd scp	$srvctl_script	\
+								root@$cfg_server_name:/etc/bash_completion.d/
+			fi
+			fake_exec_cmd "cd -"
+			cd -
+			LN
 		fi
-		fake_exec_cmd "cd -"
-		cd -
-		LN
 	fi
 
 	ssh_server "plescripts/gadgets/customize_logon.sh"
